+++ CHANGELOG.rst
@@ -25,7 +25,7 @@ New features
   exactly correct version that had been implemented.  Not all combinations are
   implemented.  So far just GGG, KKK, and NNN. (#115)
 - Added ability to read from parquet catalogs.  (#117)
-- Added optional func parameter to estimate_multi_cov. (#118)
+- Added optional func parameter to estimate_cov and estimate_multi_cov. (#118)
 - Added ability to write output files as hdf5.  (#122)
 
 Bug fixes